Set pom.xml for Windows only, enable sitting skeleton, package organization
