+++ pom.xml
@@ -13,6 +13,23 @@
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
   </properties>
+  
+  <build>
+    <pluginManagement>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.1</version>
+                <configuration>
+                    <source>1.8</source>
+                    <target>1.8</target>
+                </configuration>
+            </plugin>
+        </plugins>
+
+    </pluginManagement>
+</build>
 
   <dependencies>
     <dependency>
@@ -23,14 +40,40 @@
     </dependency>
     <dependency>
       <groupId>org.jogamp.gluegen</groupId>
-      <artifactId>gluegen-rt-main</artifactId>
+      <artifactId>gluegen-rt</artifactId>
+      <version>2.3.1</version>
+    </dependency>
+        <dependency>
+      <groupId>org.jogamp.gluegen</groupId>
+      <artifactId>gluegen-rt</artifactId>
+      <version>2.3.1</version>
+       <classifier>natives-windows-amd64</classifier>
+    </dependency>
+        <dependency>
+      <groupId>org.jogamp.gluegen</groupId>
+      <artifactId>gluegen-rt</artifactId>
+      <version>2.3.1</version>
+      <classifier>natives-windows-i586</classifier>
+    </dependency>
+    
+    
+    
+    <dependency>
+      <groupId>org.jogamp.jogl</groupId>
+      <artifactId>jogl-all</artifactId>
       <version>2.3.1</version>
     </dependency>
     <dependency>
-      <groupId>org.jogamp.jogl</groupId>
-      <artifactId>jogl-all-main</artifactId>
-      <version>2.3.1</version>
-    </dependency>
-
+            <groupId>org.jogamp.jogl</groupId>
+            <artifactId>jogl-all</artifactId>
+            <version>2.3.1</version>
+            <classifier>natives-windows-amd64</classifier>
+	</dependency>
+    <dependency>
+            <groupId>org.jogamp.jogl</groupId>
+            <artifactId>jogl-all</artifactId>
+            <version>2.3.1</version>
+            <classifier>natives-windows-i586</classifier>
+	</dependency>
     </dependencies>
 </project>
